refactor(widgets, shared): change add Carousel
