+++ eslint.config.mjs
@@ -1,6 +1,7 @@
 import { dirname } from "path";
 import { fileURLToPath } from "url";
 import { FlatCompat } from "@eslint/eslintrc";
+import reactUseEffect from "eslint-plugin-react-useeffect";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
