+++ src/shared/ui/carousel/carousel-dot.tsx
