+++ src/shared/ui/index.ts
@@ -1,3 +1,4 @@
 export { Button } from "./button";
 export { Title } from "./title";
 export { Accordion } from "./accordion";
+export { Carousel } from "./carousel/carousel";
