+++ src/widgets/coaches/ui/coaches.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import useEmblaCarousel from "embla-carousel-react";
-import { CoachCard } from "./coaches-сard";
 import { coachesData } from "../model/data";
-import { CarouselControls } from "./carousel-controls";
+import { CarouselControls } from "../../../shared/ui/carousel/carousel-controls";
 import { SelectedCoach } from "./selected-coach";
+import { Carousel } from "@/shared/ui";
 
 export function Coaches() {
   const [emblaRef, emblaApi] = useEmblaCarousel({ loop: true });
@@ -13,18 +13,7 @@ export function Coaches() {
     <section className="flex flex-col gap-6 px-5 py-10">
       <SelectedCoach />
       <div className="overflow-hidden" ref={emblaRef}>
-        <div className="flex">
-          {coachesData.map(({ name, description, src, alt }) => (
-            <div key={name} className="w-full flex-shrink-0 xl:w-1/4">
-              <CoachCard
-                name={name}
-                description={description}
-                src={src}
-                alt={alt}
-              />
-            </div>
-          ))}
-        </div>
+        <Carousel data={coachesData} />
       </div>
       <CarouselControls emblaApi={emblaApi} />
     </section>
